refactor: fix from phpstan and psalm

+++ src/Money/Money.php
@@ -6,7 +6,7 @@ namespace App\Money;
 class Money implements Expression
 {
     /**
-     * @var int $amount
+     * @var float $amount
      */
     protected $amount;
     /**
@@ -14,7 +14,7 @@ class Money implements Expression
      */
     protected $currency;
 
-    public function __construct(int $amount, string $currency)
+    public function __construct(float $amount, string $currency)
     {
         $this->amount = $amount;
         $this->currency = $currency;
@@ -63,9 +63,9 @@ class Money implements Expression
     }
 
     /**
-     * @return int
+     * @return float
      */
-    public function getAmount(): int
+    public function getAmount(): float
     {
         return $this->amount;
     }
